OurPlanner project:
Model class - contains a set of Action's (class) that represents the
domain.

TempModel class - contains a set of TempAction's (class) that represents
the changes that are added to a given model.

Planner And Model Learner - Tries to generate a plan by iterating and
improving the model using the safe/unsafe models.