OurPlanner project:
Feature 1 - Extending a model when Planning and Learning algorithms are
now filtering contradicting models, i.e. models that contain 2 or more
preconditions of the same variable.
Feature 2 - Before planning, every model is checked if he allows a
failed actions, using past knowledge. if so, the model is extended and
his children are put back on the open model list.